src/modules/roles/roles.service.ts
import { Injectable } from '@nestjs/common';
import { Repository } from 'typeorm';
import { InjectRepository } from '@nestjs/typeorm';
import { CreateRoleDto } from './dto/create-role.dto';
import { UpdateRoleDto } from './dto/update-role.dto';
import { Role } from './entities/role.entity';
import { User } from '../users/entities/user.entity';

@Injectable()
export class RolesService {
  constructor(
    @InjectRepository(Role) private roleRepository: Repository<Role>,
    @InjectRepository(User) private userRepository: Repository<User>
  ) {}

  create(createRoleDto: CreateRoleDto) {
    const user = this.userRepository.findOne({where: {id: 3}});
    const role = this.roleRepository.save(createRoleDto);
    return role;
  }

  findAll() {
    return this.roleRepository.find();
  }

  findOne(id: number) {
    return this.roleRepository.findOne({where: {id}});
  }

  update(id: number, updateRoleDto: UpdateRoleDto) {
    return this.roleRepository.update({id}, updateRoleDto);
  }

  remove(id: number) {
    return this.roleRepository.delete({id});
  }
}
